Separate util functions in tm-stats, refactor
